Remove FatalError from CfReadLine, push error handling to caller.
This changes the signature of CfReadLine from returning a true/false
to returning a ssize_t of bytes read. A return value of -1 signals
the error where previously CfReadLine terminated by FatalError.

+++ src/files_interfaces.h
@@ -10,7 +10,11 @@ int cfstat(const char *path, struct stat *buf);
 int cf_stat(char *file, struct stat *buf, Attributes attr, Promise *pp);
 int cf_lstat(char *file, struct stat *buf, Attributes attr, Promise *pp);
 int CopyRegularFile(char *source, char *dest, struct stat sstat, struct stat dstat, Attributes attr, Promise *pp, const ReportContext *report_context);
-int CfReadLine(char *buff, size_t size, FILE *fp);
+
+/**
+ * @return Number of characters read, or -1 on error.
+ */
+ssize_t CfReadLine(char *buff, size_t size, FILE *fp);
 int cf_readlink(char *sourcefile, char *linkbuf, int buffsize, Attributes attr, Promise *pp);
 
 #endif
